+++ src/instruction_set/tests/address_mode.rs
@@ -48,7 +48,7 @@ fn validate_address_mode_with_operands_into_bytes_returns_correct_endian_bytes()
 
 #[test]
 fn validate_address_mode_maps_to_the_correct_size_of() {
-    let address_sizing: Vec<usize> = vec![
+    let address_sizing: Vec<u16> = vec![
         AddressMode::Absolute(0x8008),
         AddressMode::AbsoluteIndexedWithX(0x8008),
         AddressMode::AbsoluteIndexedWithY(0x8008),
